Status effect source reference; ActionCast battle log work; Cleaning compiler warnings;

+++ src/servers/Server_Zone/Actor/CalcBattle.h
@@ -12,7 +12,7 @@ namespace Entity {
    {
    public:
 
-      static uint32_t calculateBaseStat( PlayerPtr pPlayer );
+      static float calculateBaseStat( PlayerPtr pPlayer );
       static uint32_t calculateMaxMp( PlayerPtr pPlayer );
       static uint32_t calculateMaxHp( PlayerPtr pPlayer );
       static uint32_t calculateHealValue( PlayerPtr pPlayer, uint32_t potency );
